Wait for URLProtocol to complete the URLRequest
after a task has been canceled to avoid a test leak where the URLProtocol would be start the request after the test finishes. This happens because cancelling a URLSessionDataTask won't immediately cancel the URLProtocol from receiving that request. So if we don't wait for it, there's a chance the URLProtocol request will run while another test is running and influence its result.